requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
Faker==29.0.0
inflection==0.5.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
